pyproject.toml
[project]
name = "ai-analytics-insight"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "fastapi>=0.128.0",
    "langchain>=1.2.3",
    "langchain-community>=0.4.1",
    "langchain-core>=1.2.7",
    "langchain-groq>=1.1.1",
    "langsmith>=0.6.2",
    "mysql-connector-python>=9.5.0",
    "pandas>=3.0.0",
    "pymysql>=1.1.2",
    "python-dotenv>=1.2.1",
    "sqlalchemy>=2.0.45",
    "typedict>=0.1.0",
    "uvicorn>=0.40.0",
]
